Trasfere acesso à api para a primeira view e inicia processo de exclusão da terceira tela
