+++ Finx - Agamenon/AppSurdo/Models/RandomFact.swift
