+++ Finx - Agamenon/AppSurdo/Views/Storyboards/Main.storyboard
@@ -37,7 +37,7 @@
                                 <gestureRecognizers/>
                                 <state key="normal" image="PlayButton"/>
                                 <connections>
-                                    <segue destination="BV1-FR-VrT" kind="presentation" id="zvh-qw-8jW"/>
+                                    <segue destination="BV1-FR-VrT" kind="presentation" identifier="ARSegue" id="zvh-qw-8jW"/>
                                 </connections>
                             </button>
                             <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="a8C-ka-wcW">
@@ -94,7 +94,7 @@
                                     <constraint firstAttribute="width" constant="84" id="Vf8-Ch-61z"/>
                                     <constraint firstAttribute="height" constant="84" id="WkK-Ol-Z56"/>
                                 </constraints>
-                                <state key="normal" image="Card"/>
+                                <state key="normal" backgroundImage="Card"/>
                             </button>
                             <button opaque="NO" contentMode="scaleToFill" misplaced="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="EK8-MR-Ylu">
                                 <rect key="frame" x="271" y="674" width="84" height="84"/>
@@ -102,7 +102,7 @@
                                     <constraint firstAttribute="height" constant="84" id="5ZY-1N-mDm"/>
                                     <constraint firstAttribute="width" constant="84" id="R4r-T7-jHi"/>
                                 </constraints>
-                                <state key="normal" image="Card"/>
+                                <state key="normal" backgroundImage="Card"/>
                             </button>
                             <view hidden="YES" contentMode="scaleToFill" ambiguous="YES" translatesAutoresizingMaskIntoConstraints="NO" id="QnH-B7-yh0">
                                 <rect key="frame" x="50.328163553088189" y="288.82845309407412" width="309.00000000000028" height="266.99999999999983"/>
@@ -147,7 +147,7 @@
                                     <constraint firstAttribute="height" constant="84" id="oKA-gO-RaA"/>
                                     <constraint firstAttribute="width" constant="84" id="wEE-zW-FWc"/>
                                 </constraints>
-                                <state key="normal" image="Card"/>
+                                <state key="normal" backgroundImage="Card"/>
                             </button>
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="fQZ-KI-GVf"/>
